Removed process_exception(). Improved docs.
This fork was created because the implementation of process_exception() breaks
the ability to retrieve the current request during exception handling
(e.g. rendering a 404 or 500 page).

It never needed to be implemented in the first place, because the way that
Django 1.9 and below process exceptions means that either process_response()
will be called, or neither process_response() nor process_exception() will be
called. Django 1.10+, with the MIDDLEWARE setting, guarantees that
process_response() will always be called, so process_exception() is not needed
there, either.